+++ src/repository/TipoContatoRepository.java
@@ -3,6 +3,7 @@ package repository;
 import model.Segmento;
 import model.TipoContato;
 
+import javax.swing.*;
 import java.sql.*;
 import java.util.ArrayList;
 import java.util.List;
@@ -23,7 +24,9 @@ public class TipoContatoRepository {
         stmt.setString(1, tipoContato.getNome());
 
         int i = stmt.executeUpdate();
-        System.out.println(i + " linhas inseridas");
+        if (i > 0){
+            JOptionPane.showMessageDialog(null, "Tipo de Contato inserido!");
+        }
         connection.close();
     }
 
@@ -72,16 +75,37 @@ public class TipoContatoRepository {
         stmt.setString(2, tipoContato.getNome());
 
         int i = stmt.executeUpdate();
-        System.out.println(i + " linhas atualizadas");
+        if (i > 0){
+            JOptionPane.showMessageDialog(null, "Tipo de Contato atualizado!");
+        }
         connection.close();
     }
 
     public void delete(TipoContato tipoContato) throws SQLException, ClassNotFoundException {
         Connection connection = getConnection();
-        PreparedStatement stmt = connection.prepareStatement("DELETE FROM TipoContato" +
-                " WHERE id = ?");
-        stmt.setInt(1, tipoContato.getId().intValue());
-        stmt.executeUpdate();
-        connection.close();
+        int count = 0;
+
+        PreparedStatement stmt2 = connection.prepareStatement("select count(*) from contato where id = ?");
+        stmt2.setInt(1, tipoContato.getId().intValue());
+        ResultSet resultSet = stmt2.executeQuery();
+
+        while(resultSet.next()){
+            count = resultSet.getInt(1);
+            if (count > 0){
+                JOptionPane.showMessageDialog(null, "Não é possível excluir o tipo de contato pois ele está vinculada à algum contato!");
+            }
+        }
+
+        if (count ==0) {
+
+            PreparedStatement stmt = connection.prepareStatement("DELETE FROM TipoContato" +
+                    " WHERE id = ?");
+            stmt.setInt(1, tipoContato.getId().intValue());
+            int i = stmt.executeUpdate();
+            if (i > 0){
+                JOptionPane.showMessageDialog(null, "Tipo de contato excluído!");
+            }
+            connection.close();
+        }
     }
 }
